% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{pca_projection_R}
\alias{pca_projection_R}
\title{Compute the projection}
\usage{
pca_projection_R(C, L)
}
\arguments{
\item{C}{description}

\item{L}{description}
}
\description{
Compute the projection
}

