% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{sqdist_R}
\alias{sqdist_R}
\title{perform PCA projection
solve the problem size(C) = NxN, size(W) = NxL
max_W trace( W' C W ) : W' W = I}
\usage{
sqdist_R(a, b)
}
\arguments{
\item{a}{a matrix with D x N dimension}

\item{b}{a matrix with D x N dimension}
}
\value{
a numeric value for the different between a and b
}
\description{
perform PCA projection
solve the problem size(C) = NxN, size(W) = NxL
max_W trace( W' C W ) : W' W = I
}

