% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{get_major_eigenvalue}
\alias{get_major_eigenvalue}
\title{Get the top L eigenvalues}
\usage{
get_major_eigenvalue(C, L)
}
\arguments{
\item{C}{data matrix used for eigendecomposition}

\item{L}{number for the top eigenvalues}
}
\description{
Get the top L eigenvalues
}
