% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{pca_projection_R}
\alias{pca_projection_R}
\title{Compute the PCA projection}
\usage{
pca_projection_R(C, L)
}
\arguments{
\item{C}{data matrix used for PCA projection}

\item{L}{number for the top principal components}
}
\description{
Compute the PCA projection
}
