% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{sqdist_R}
\alias{sqdist_R}
\title{calculate the square distance between a, b}
\usage{
sqdist_R(a, b)
}
\arguments{
\item{a}{a matrix with \eqn{D \times N} dimension}

\item{b}{a matrix with \eqn{D \times N} dimension}
}
\value{
a numeric value for the different between a and b
}
\description{
calculate the square distance between a, b
}
