% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{sqdist_R}
\alias{sqdist_R}
\title{perform PCA projection
solve the problem size(C) = \eqn{N \times N}, size(W) = \eqn{N \times L}
\eqn{max_\mathbf{W} trace(\mathbf{W}' \mathbf{C W} }) : \eqn{\mathbf{W}' \mathbf{W} = \mathbf{I}}}
\usage{
sqdist_R(a, b)
}
\arguments{
\item{a}{a matrix with \eqn{D \times N} dimension}

\item{b}{a matrix with \eqn{D \times N} dimension}
}
\value{
a numeric value for the different between a and b
}
\description{
perform PCA projection
solve the problem size(C) = \eqn{N \times N}, size(W) = \eqn{N \times L}
\eqn{max_\mathbf{W} trace(\mathbf{W}' \mathbf{C W} }) : \eqn{\mathbf{W}' \mathbf{W} = \mathbf{I}}
}

